Extract getSecondaryUrlFromPrimary to platform-specific files
