+++ sdk/tables/data-tables/src/getSecondaryUrlFromPrimary-react-native.mts
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export { getSecondaryUrlFromPrimary } from "./getSecondaryUrlFromPrimary-browser.mjs";
